pyproject.toml
[project]
name = "agent-max"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.34.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "tavily>=1.1.0",
]
